+++ public/CHANGELOG.md
@@ -19,6 +19,24 @@ The format is loosely based on [Keep a Changelog](https://keepachangelog.com/en/
 - Optional filters in upgrade menu
 - More CRT effects
 
+## 0.0.2 - 23.03.2025
+
+### Added
+
+- Counter for complete rendered frames
+- New Changelog popup when loading the game
+- New Changelog button in the settings menu
+
+### Changed
+
+- Made the color shift upon completion of a frame more pronounced
+- Moved the changelog to the public folder (to be included in the build)
+
+### Fixed
+
+- Fixed a bug where the canvas was not reset correctly upon completion of a frame
+- Fixed styling issues with the settings and upgrades menus
+
 ## 0.0.1 - 23.03.2025
 
 ### Added